[owners] Move ownership logic into `ownership` submodule
