+++ owners/test/owners_bot.test.js
@@ -19,8 +19,8 @@ const {GitHub, PullRequest, Review, Team} = require('../src/api/github');
 const LocalRepository = require('../src/repo/local_repo');
 const {OwnersBot} = require('../src/owners_bot');
 const {OwnersNotifier} = require('../src/notifier');
-const {OwnersTree} = require('../src/owners_tree');
-const {CheckRun, CheckRunState, OwnersCheck} = require('../src/owners_check');
+const OwnersTree = require('../src/ownership/tree');
+const {CheckRun, CheckRunState, OwnersCheck} = require('../src/ownership/owners_check');
 
 describe('owners bot', () => {
   const silentLogger = {
